chore: LuckyCard 리팩토링
